Move gen utils to separate gencommons
