+++ pkg/internal/gencommons/util.go
@@ -1,4 +1,4 @@
-package gen
+package gencommons
 
 import (
 	"regexp"
