+++ pkg/internal/gencommons/util_test.go
@@ -1,4 +1,4 @@
-package gen
+package gencommons
 
 import (
 	"fmt"
